pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raytool"
version = "2.1.0"
description = "Ray 集群任务管理命令行工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "InquirerPy>=0.3",
    "pyyaml>=6.0",
]

[project.scripts]
raytool = "raytool.cli:cli"

[project.urls]
Homepage = "https://github.com/ZhijianZhou/Qytool"

[tool.setuptools.package-dir]
raytool = "."
"raytool.commands" = "commands"
"raytool.utils" = "utils"

[tool.setuptools]
packages = ["raytool", "raytool.commands", "raytool.utils"]
